Summon all Antlers nodes below a common root element
Currently, all Antlers nodes are siblings without a common parent node. To get a better structure create a parent node and mount all Antlers nodes as children of that node. 

This helps to traverse the AST tree later on and is needed for implementing the formatting feature.

I called the parent node Tines and a Antlers "thing" now tine, as this is what the spikes or branches of an Antlers are called. Normally I would have called the whole construct a "tag" but Antlers already has Tags. Node was fine too, but the AST tree uses that word as well. So Tines it is.