Move the lexing of PHP nodes to Antlers state
We changed the initial state to detect everything up to the open double curly braces `{{`. Everything before that (except a @) is considered HTML. We put the `{{` chars back into the game and handle them in the more dedicated ANTLERS state. Therefore we need to move the PHP and Comment structures into this state.